requirements.txt
tensorflow
pandas
streamlit
sqlalchemy
psycopg2-binary
python-dotenv
pillow
scikit-learn